Added pause/resume functionality
Added pause button to top right corner -> enables pause-panel, with button to resume game.
Adjusted restart funtionality, adding a "Game Over" panel with button.

Fixed an issue, where the "restart" button could not be clicked, upon the player dying. The issue was that the EventSystem object was a child to the player object (which gets destroyed upon player-death -> disabling the EventSystem)